% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queuenet_object.R
\name{construct_queuenets_object}
\alias{construct_queuenets_object}
\title{Define an queuenets object}
\usage{
construct_queuenets_object(nodes = 4)
}
\arguments{
\item{nodes}{Integer. Number of nodes.}
}
\value{
A list structure of class \code{queuenets}.
}
\description{
\lifecycle{experimental}
COnstructs an queuenets object.
}
