% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queuenet_object.R
\name{is_queuenets_object}
\alias{is_queuenets_object}
\title{Checks if network is correctly defined}
\usage{
is_queuenets_object(object, detail = FALSE)
}
\arguments{
\item{object}{Object to be checked.}

\item{detail}{Logical. If TRUE, details are returned}
}
\value{
Logical or list for detail = TRUE. TRUE, if conditions are fulfilled.
}
\description{
\lifecycle{experimental}
Checks, if the object fulfills the conditions defined in \link{code{construct_queuenets_object()}}.
}
